+++ packages/core/src/Exceptions/DelirException.ts
