+++ packages/core/src/Exceptions/index.ts
@@ -1,8 +1,8 @@
-export { default as DelirException } from './delir-exception'
-export { default as PluginLoadFailException } from './plugin-load-fail-exception'
-export { default as UnknownPluginReferenceException } from './unknown-plugin-reference-exception'
-export { default as PluginAssertionFailedException } from './plugin-assertion-failed-exception'
-export { default as RenderingFailedException } from './rendering-failed-exception'
-export { default as RenderingAbortedException } from './rendering-aborted-exception'
+export { default as DelirException } from './DelirException'
+export { default as PluginLoadFailException } from './PluginLoadFailException'
+export { default as UnknownPluginReferenceException } from './UnknownPluginReferenceException'
+export { default as PluginAssertionFailedException } from './PlugiinAssertionFailedException'
+export { default as RenderingFailedException } from './RenderingFailedException'
+export { default as RenderingAbortedException } from './RenderingAbortedException'
 export { default as EffectPluginMissingException } from './EffectPluginMissingException'
 export { UserCodeException } from './UserCodeException'
